// studio/src/utils/zitadel/types.ts
export interface SessionData {
  sessionId: string;
  sessionToken: string;
  token: string;
}

export interface UserDetails {
  session: {
    factors: {
      user: {
        id: string;
      };
    };
  };
}

export interface AuthMethods {
  authMethodTypes: string[];
}

export interface FinalizeResult {
  callbackUrl?: string;
}